+++ packages/browser-wallet/src/popup/constants/urls.ts
@@ -1,0 +1,8 @@
+const urls = {
+    termsAndConditions: 'https://developer.concordium.software/en/mainnet/net/resources/terms-and-conditions-bw.html',
+    website: 'https://www.concordium.com',
+    supportWebsite: 'https://support.concordium.software',
+    documentationWebsite: 'https://developer.concordium.software/en/mainnet/index.html',
+};
+
+export default urls;