+++ packages/browser-wallet/src/popup/pages/Setup/i18n/da.ts
@@ -5,6 +5,11 @@ const t: typeof en = {
         welcome: "Velkommen til Concordium's officielle browser extension wallet",
         description:
             'På de følgende sider vil du blive guidet igennem at oprette en ny konto, eller at genskabe allerede eksisterende konti.',
+        form: {
+            termsAndConditionsRequired: 'Det er påkrævet at læse og at acceptere vilkår og betingelser',
+            termsAndConditionsDescription: 'Jeg har læst og accepterer',
+            termsAndConditionsLinkDescription: 'Vilkår og Betingelser',
+        },
     },
     setupPasscode: {
         title: 'Vælg kodeord',