+++ packages/browser-wallet/src/popup/shared/ExternalLink/ExternalLink.tsx
@@ -1,0 +1,9 @@
+import React from 'react';
+
+export default function ExternalLink({ path, label }: { path: string; label: string }) {
+    return (
+        <a href={`${path}`} target="_blank" rel="noreferrer">
+            {label}
+        </a>
+    );
+}